Prep to separate element definitions to separate file.

+++ include/element_dof.hpp
@@ -3,10 +3,11 @@
 
 #include "../Eigen/Core"
 
-typedef Eigen::Matrix<float, 4, 4> Matrix4f;
-typedef Eigen::Matrix<float, 4, 1> Vector4f;
+typedef Eigen::Matrix<double, 4, 4> Matrix4d;
+typedef Eigen::Matrix<double, 4, 1> Vector4d;
 
-Vector4f getElementDof(float, float);
-Matrix4f getElementK(float, float);
+Vector4d getElementDof(double, double);
+Matrix4d getElementK(double, double);
+// Vector4f getElementF(float, float);
 
 #endif
